Separates DisplayElement classes into their own files
